Update meter_page : Seperated meter_costview widget
